refactor: 🔨 convert into library
