--- entities/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "entities"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "entities"
-path = "src/lib.rs"
-
-[dependencies]
-# common
-eyre = { workspace = true, default-features = false }
-lool = { workspace = true, features = [ "cli", "cli.stylize", "macros", "logger" ] }
-
-# external
-sea-orm = { version = "0.12.15", features = [
-    "runtime-tokio-native-tls",
-    "sqlx-sqlite",
-    "macros",
-] }
