--- entities/src/orm/market.rs
@@ -1,33 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "market")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: String,
-    pub short_name: String,
-    pub full_name: String,
-    pub opens_from: Option<u32>,
-    pub opens_till: Option<u32>,
-    pub open_time: Option<String>,
-    pub close_time: Option<String>,
-    pub pre_market_offset: Option<u32>,
-    pub post_market_offset: Option<u32>,
-    pub time_zone_offset: Option<String>,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(has_many = "super::ticker::Entity")]
-    Ticker,
-}
-
-impl Related<super::ticker::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Ticker.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
