--- gateway/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "gateway"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "gateway"
-path = "src/lib.rs"
-
-[dependencies]
-entities = { path = "../entities" }
-eyre = { workspace = true, default-features = false }
-
