--- grpc/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "grpc"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "grpc"
-path = "src/lib.rs"
-
-[dependencies]
-# common
-lool = { workspace = true, features = ["logger"]}
-eyre = { workspace = true, default-features = false }
-
-# internal
-entities = { path = "../entities" }
-
-# external
-prost = "0.12.3"
-tonic = "0.11.0"
-
-[build-dependencies]
-tonic-build = "0.11.0"
-
