+++ lib/entities/migration/mod.rs
@@ -1,0 +1,2 @@
+mod m20220101_000001_create_table_market;
+mod m20240325_200049_create_table_ticker;
