+++ lib/lib.rs
@@ -1,0 +1,4 @@
+pub mod entities;
+pub mod grpc;
+pub mod rustlers;
+pub mod socket;
