+++ lib/socket/mod.rs
@@ -1,0 +1,1 @@
+// TODO: websocket gateway
