--- migration/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "migration"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "migration"
-path = "src/lib.rs"
-
-[dependencies]
-async-std = { version = "1", features = ["attributes", "tokio1"] }
-entities = { path = "../entities" }
-
-[dependencies.sea-orm-migration]
-version = "0.12.15"
-features = [
-  "runtime-tokio-native-tls",
-  "sqlx-sqlite",
-]
