--- rustlers/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "rustlers"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "rustlers"
-path = "src/lib.rs"
-
-[dependencies]
-# workspace
-eyre = { workspace = true, default-features = false }
-tokio-tungstenite = { workspace = true }
-tokio = { workspace = true }
-lool = { workspace = true, features = [
-    "logger",
-    "sched.tokio",
-    "sched.rule-recurrence",
-    "macros",
-] }
-
-# internal
-entities = { path = "../entities" }
-
-# external
-chrono = "0.4.37"
-async-trait = "0.1.79"
-getset = "0.1.2"
-
-[build-dependencies]
-tonic-build = "0.11.0"
