--- rustlers/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod rustlerjar;
-pub mod rustlers;
-pub mod svc;
